Add account provider abstraction
